marine_conditions/requirements.txt
supabase>=2.5.0
requests>=2.32.0
python-dateutil>=2.9.0
